requirements.txt
Flask==3.0.2
python-gitlab==4.4.0